Marshal AMQP message data
The AMQP client does not seem to work with the latest tmm1-amqp gem.
The message placed on the queue was the result of calling
value#to_s. When the message was pulled off the queue it was the
string representation of the value hash.

This patch marshals the message Hash before placing publishing it
and then unmarshals when it picks it up.

Signed-off-by: Rany Keddo <rany@playtype.net>